requirements/production.txt
Django==1.11.27
Pillow==4.0.0
boto==2.36.0
psycopg2
factory-boy==2.1.2
Unipath==1.1
django-braces
django-spurl
sorl-thumbnail==12.4.1
django-uuslug==1.0.0
django-model-utils==1.5.0
django-aggregate-if==0.3.1
django-kombu==0.9.4
django-extra-views==0.8.0
django-skd-tools==0.0.10
django-storages==1.5.1
pysolr==3.1.0
python-dateutil==2.2
django-constance==2.4.0
pytz==2016.10
django-formtools==2.0
django-picklefield==0.3.2
user-agents==1.1.0
pyfcm==1.2.1
requests==2.13.0
requests-oauthlib==0.8.0
mock==2.0.0
beautifulsoup4==4.7.1
django-file-resubmit==0.5.2
django-redis==4.10.0
